# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CCMS API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/social-sessions": {
      "get": {
        "tags": [
          "social"
        ],
        "summary": "List Sessions",
        "operationId": "list_sessions_api_v1_social_sessions_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SocialSessionOut"
                  },
                  "title": "Response List Sessions Api V1 Social Sessions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "social"
        ],
        "summary": "Create Session",
        "operationId": "create_session_api_v1_social_sessions_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SocialSessionCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SocialSessionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/social-sessions/{session_id}": {
      "get": {
        "tags": [
          "social"
        ],
        "summary": "Get Session Row",
        "operationId": "get_session_row_api_v1_social_sessions__session_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SocialSessionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "social"
        ],
        "summary": "Cancel Session",
        "operationId": "cancel_session_api_v1_social_sessions__session_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SocialSessionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/social-sessions/{session_id}/participants": {
      "get": {
        "tags": [
          "social"
        ],
        "summary": "Participants",
        "description": "Staff only: the roster names other members (SRS 3.2.6, S-15).",
        "operationId": "participants_api_v1_social_sessions__session_id__participants_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SocialParticipantOut"
                  },
                  "title": "Response Participants Api V1 Social Sessions  Session Id  Participants Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/social-sessions/{session_id}/join": {
      "post": {
        "tags": [
          "social"
        ],
        "summary": "Join",
        "operationId": "join_api_v1_social_sessions__session_id__join_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SocialJoin"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SocialParticipantOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/social-sessions/{session_id}/leave": {
      "post": {
        "tags": [
          "social"
        ],
        "summary": "Leave",
        "operationId": "leave_api_v1_social_sessions__session_id__leave_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SocialLeave"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SocialJoin": {
        "properties": {
          "member_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Member Id"
          },
          "guest_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Guest Name"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "SocialJoin"
      },
      "SocialLeave": {
        "properties": {
          "member_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Member Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "SocialLeave"
      },
      "SocialParticipantOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "session_id": {
            "type": "integer",
            "title": "Session Id"
          },
          "member_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Member Id"
          },
          "guest_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Guest Name"
          },
          "fee_paise": {
            "type": "integer",
            "title": "Fee Paise"
          },
          "status": {
            "$ref": "#/components/schemas/SocialParticipantStatus"
          }
        },
        "type": "object",
        "required": [
          "id",
          "session_id",
          "member_id",
          "guest_name",
          "fee_paise",
          "status"
        ],
        "title": "SocialParticipantOut"
      },
      "SocialParticipantStatus": {
        "type": "string",
        "enum": [
          "JOINED",
          "LEFT"
        ],
        "title": "SocialParticipantStatus"
      },
      "SocialSessionCreate": {
        "properties": {
          "court_id": {
            "type": "integer",
            "title": "Court Id"
          },
          "title": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Title"
          },
          "start_at": {
            "type": "string",
            "format": "date-time",
            "title": "Start At"
          },
          "end_at": {
            "type": "string",
            "format": "date-time",
            "title": "End At"
          },
          "capacity": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Capacity"
          },
          "fee_paise": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Fee Paise",
            "default": 0
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "court_id",
          "title",
          "start_at",
          "end_at",
          "capacity"
        ],
        "title": "SocialSessionCreate"
      },
      "SocialSessionOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "court_id": {
            "type": "integer",
            "title": "Court Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "start_at": {
            "type": "string",
            "format": "date-time",
            "title": "Start At"
          },
          "end_at": {
            "type": "string",
            "format": "date-time",
            "title": "End At"
          },
          "capacity": {
            "type": "integer",
            "title": "Capacity"
          },
          "joined_count": {
            "type": "integer",
            "title": "Joined Count"
          },
          "fee_paise": {
            "type": "integer",
            "title": "Fee Paise"
          },
          "status": {
            "$ref": "#/components/schemas/SocialSessionStatus"
          }
        },
        "type": "object",
        "required": [
          "id",
          "court_id",
          "title",
          "start_at",
          "end_at",
          "capacity",
          "joined_count",
          "fee_paise",
          "status"
        ],
        "title": "SocialSessionOut"
      },
      "SocialSessionStatus": {
        "type": "string",
        "enum": [
          "OPEN",
          "CANCELLED"
        ],
        "title": "SocialSessionStatus"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}